chore: update storybook config to latest setup, resolving ts imports within mdx files chore: rename to .module.scss, so that storybook default config can pick up and resolve
