+++ .storybook/preview.js
@@ -1,0 +1,1 @@
+import './../src/styles/main.scss';
