+++ package.json
@@ -58,6 +58,7 @@
     "react-redux": "^7.1.1",
     "react-router": "^5.1.2",
     "react-router-dom": "^5.1.2",
+    "react-scripts": "^3.4.0",
     "react-tabs": "^2.3.0",
     "react-tooltip": "^3.11.3",
     "redux": "^4.0.0",
@@ -77,10 +78,11 @@
     "@babel/preset-env": "^7.6.2",
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-typescript": "7.6.0",
-    "@storybook/addon-actions": "^5.2.1",
-    "@storybook/addon-docs": "^5.2.1",
-    "@storybook/addons": "^5.2.1",
-    "@storybook/react": "^5.2.1",
+    "@storybook/addon-actions": "^5.3.14",
+    "@storybook/addon-docs": "^5.3.14",
+    "@storybook/addons": "^5.3.14",
+    "@storybook/preset-create-react-app": "^2.0.0",
+    "@storybook/react": "^5.3.14",
     "@types/debug": "4.1.5",
     "@types/enzyme": "3.10.3",
     "@types/jest": "24.0.18",
