+++ src/components/Helpers/ButtonSolid/ButtonSolid.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Props, Preview } from "@storybook/addon-docs/blocks";
-import { ButtonSolid } from "./ButtonSolid.tsx";
-import { SvgIcon, SvgIconEdit } from "./../SvgIcon/SvgIcon.tsx";
+import { ButtonSolid } from "./ButtonSolid";
+import { SvgIcon, SvgIconEdit } from "./../SvgIcon";
 
 <Meta title="MDX|ButtonSolid" component={ButtonSolid} />
