+++ src/components/Helpers/LabelBordered/LabelBordered.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Props, Preview } from "@storybook/addon-docs/blocks";
-import { LabelBordered } from "./LabelBordered.tsx";
+import { LabelBordered } from "./LabelBordered";
 
 <Meta title="MDX|LabelBordered" component={LabelBordered} />
