+++ src/components/Helpers/SvgIcon/SvgIcon.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Props, Preview } from "@storybook/addon-docs/blocks";
-import { SvgIcon, SvgIconInfo, SvgIconChecked, SvgIconChervonLeft, SvgIconEdit } from "./SvgIcon.tsx";
+import { SvgIcon, SvgIconInfo, SvgIconChecked, SvgIconChervonLeft, SvgIconEdit } from "./SvgIcon";
 
 <Meta title="MDX|SvgIcon" component={SvgIcon} />
