+++ src/components/PrintWatermark.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Preview, Props, Description, Story } from "@storybook/addon-docs/blocks";
 import { PrintWatermark } from "./PrintWatermark";
 
-<Meta title="MDX|PrintWatermark" component={PrintWatermark} />
+<Meta title="MDX|Others" component={PrintWatermark} />
 
 # PrintWatermark component
 
@@ -17,7 +17,7 @@ With `MDX` we can define a story for `PrintWatermark` right in the middle of our
 markdown documentation.
 
 <Preview withSource>
-  <Story name="basic example">
+  <Story name="PrintWatermark">
     <PrintWatermark />
   </Story>
 </Preview>
